pyproject.toml
# SPDX-FileCopyrightText: 2021 The SMW Music Python Project Authors
# <https://github.com/com-posers-pit/smw_music/blob/develop/AUTHORS.rst>
#
# SPDX-License-Identifier: CC0-1.0

[tool.bandit]
exclude_dirs = ["tests"]

[tool.black]
line-length = 79

[tool.isort]
import_heading_stdlib = "Standard library imports"
import_heading_thirdparty = "Library imports"
import_heading_firstparty = "Package imports"
line_length = 79
profile = "black"
filter_files = true
known_first_party = ["smw_music"]
known_third_party = ["music21"]
skip_glob = ["smw_music/doc/"]

[tool.poetry]
name = "smw_music"
version = "0.3.2"
description = "Tools for working with SMW Music"
authors = ["Thomas A. Werne <werneta@gmail.com>"]
license = "AGPL-3.0-only"
readme = "README.rst"
repository = "http://github.com/com-posers-pit/smw_music"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: Multimedia :: Sound/Audio :: Conversion",
    "Topic :: Software Development :: Libraries",
]

[tool.poetry.dependencies]
python = "^3.10,<3.12"
music21 = "^8.1.0"
Mako = "^1.1.6"
Flask = {version = "^2.0.2", optional = true}
PyQt6 = "^6.2.3"
scipy = "^1.8.0"
pyyaml = "^6.0"
types-pyyaml = "^6.0.12.2"
pyqtgraph = "^0.13.1"

[tool.poetry.group.midi]
optional = true

[tool.poetry.group.midi.dependencies]
mido = "^1.2.10"
python-rtmidi = {git = "https://github.com/SpotlightKid/python-rtmidi.git", rev = "eb16ab3268b29b94cd2baa6bfc777f5cf5f908ba"}

[tool.poetry.extras]
webserver = ["flask"]

[tool.poetry.scripts]
smw_music_filttool = "smw_music.scripts.filttool:main"
smw_music_xml_to_mml = "smw_music.scripts.convert:main"
beer = "smw_music.scripts.dashboard:main"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/com-posers-pit/smw_music/issues"

[tool.poetry.group.dev.dependencies]
Sphinx = "^5.0.0"
sphinx-rtd-theme = "^1.0.0"
sphinxcontrib-mermaid = "^0.7.1"
tox = "^4.0.0"
pre-commit = "^2.16.0"
pyinstaller = "^5.7.0"
pyqt6-stubs = {git = "https://github.com/TilmanK/PyQt6-stubs", rev = "ab869638f92b7c5264d7d169093233f15716f1cc"}

[tool.pydocstyle]
convention = "numpy"
add-ignore = "D105"

[tool.pylint.FORMAT]
good-names = "e,n,rv"

[tool.pylint.MASTER]
extension-pkg-allow-list = "PyQt6"

[tool.pylint.SIMILARITIES]
ignore-imports = "yes"

[build-system]
requires = ["poetry-core>=1.3.2"]
build-backend = "poetry.core.masonry.api"
